+++ CHANGELOG.md
@@ -19,6 +19,7 @@ nav_order: 1
   are retained, but the attribute is no longer populated.
 - Migrate `aiven_kafka_native_acl` resource to the Plugin Framework.
 - Migrate `aiven_kafka_user` resource and data source to the Plugin Framework.
+- Migrate `aiven_valkey_user` resource and data source to the Plugin Framework.
 - Fix `aiven_kafka_topic` showing persistent configuration changes when `config` values match the API defaults.
 - Migrate `aiven_organization_vpc` resource and data source to the Plugin Framework and mark them as no longer beta.
 - Migrate `aiven_aws_privatelink` resource and data source to the Plugin Framework.